+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Summary Changelog:
 - Update requirements to Node.js 14 LTS (14.15.0)
 - Ensure all is good as before
 - Update errors raised (by checkers) when no variable name is given
+- Add some functions to verify and to check/ensure if code is running as an EcmaScript module (ESM) 
+  but in a simple way
 - Simplify Tap tests execution and small improvements in code coverage
 
 ## [0.3.0](https://github.com/smartiniOnGitHub/check-runtime-env.js/releases/tag/0.3.0) (2022-03-26)